Fixed execution slow down and output bug.
After running each generation, a "cleanup" is performed where some
data is written to a file (which slows the tests down significantly
rather than writing it all at the end).

In addition, after the cleanup, all individuals in a generation are
lost, so the next iteration the species ID for each generation is
recalculated and then lost as well (since each generation except
the last only keeps the best individual). This resulted in an output
log with a value of "one" for the SpeciesCount attribute of all
generations except the last, which is not useful. :)